+++ content/footer.php
@@ -1,7 +1,11 @@
+			<div id="scroll-top">
+				<i class="fas fa-chevron-up"></i>
+			</div>
 		</main>
 		<footer class="footer">
 		
 		</footer>
+		<?php getScript('script.js'); ?>
 		<?php getThemeScript('script.js'); ?>
 	</body>
 </html>