+++ includes/globals.php
@@ -5,7 +5,7 @@
  */
 
 // Current CMS version
-const VERSION = '2.2.3';
+const VERSION = '2.2.4';
 
 /**
  * Display the copyright information on the admin dashboard.